# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/travels": {
            "get": {
                "tags": [
                    "Travel-Destination"
                ],
                "summary": "Show all TravelDestinations",
                "operationId": "378593ed20b2ac6d06668b756440158d",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Алатын-Арашан"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "1-altyn-arashan"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "Test.png"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Test test hello this is test description"
                                                    },
                                                    "destination": {
                                                        "type": "string",
                                                        "example": "Кыргызстан"
                                                    },
                                                    "destination_slug": {
                                                        "type": "string",
                                                        "example": "1-kyrgyzstan"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Travel-Destination Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/travels/{travelDestination}": {
            "get": {
                "tags": [
                    "Travel-Destination"
                ],
                "summary": "Show TravelDestination",
                "operationId": "a8b6d3de06b6cb059dc716762f160e92",
                "parameters": [
                    {
                        "name": "travelDestination",
                        "in": "path",
                        "description": "ID travelDestination",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Алатын-Арашан"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "1-altyn-arashan"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "Test.png"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Test test hello this is test description"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Кыргызстан"
                                                },
                                                "destination_slug": {
                                                    "type": "string",
                                                    "example": "1-kyrgyzstan"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Travel-Destination Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}